fix login screen design
