+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
 
     defaultConfig {
         applicationId "com.example.shows_saratedd"
-        minSdk 31
+        minSdk 26
         targetSdk 32
         versionCode 1
         versionName "1.0"
