+++ app/src/main/AndroidManifest.xml
@@ -14,13 +14,10 @@
         android:theme="@style/Theme.Showssaratedd"
         tools:targetApi="31" >
         <activity
-            android:name=".LoginActivity"
-            android:exported="false" />
-        <activity
             android:name=".WelcomeActivity"
             android:exported="false"/>
         <activity
-            android:name=".MainActivity"
+            android:name=".LoginActivity"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
