+++ app/src/main/res/color/selector_text_input_border.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/white" android:state_hovered="true"/>
+    <item android:color="@color/white" android:state_focused="true"/>
+    <item android:color="@color/white" />
+</selector>
