+++ app/src/main/res/layout/activity_welcome.xml
@@ -21,7 +21,7 @@
         android:id="@+id/logedInUser"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="text"
+        android:text=""
         android:textColor="#808080"
         android:textSize="20dp"
         android:textStyle="bold"
