+++ app/src/main/res/values/styles.xml
@@ -2,11 +2,12 @@
 <resources>
     <style name="customOutlinedBox" parent="Widget.MaterialComponents.TextInputLayout.OutlinedBox">
         <item name="android:textColor">@color/white</item>
-        <item name="boxStrokeColor">@color/white</item>
+        <item name="boxStrokeColor">@color/selector_text_input_border</item>
         <item name="android:strokeColor">@color/white</item>
         <item name="android:textColorHint">@color/white</item>
         <item name="android:textCursorDrawable">@null</item>
-
+        <item name="colorControlActivated">@color/white</item>
+        <item name="hintTextColor">@color/white</item>
     </style>
 
 </resources>
